Merge 3 pull requests from aureq.
Mozilla seems to have places link inside the h2 and h3 headers to help
with easier references. They enforce the link color to a quite dark
color making it quite difficult to see over the already dark background.

This small patch addresses the issue by apply the same grey colors to
links present in h2 and h3 headers.

Mozilla has added a breadcrumb at the top of the page and set a class
named `breadcrumb-locale-container` to define its background color to a
very light color.

This patch addresses the breadcrumb background color and reverts it to
the `page-header` class as defined in this user-style.

Mozilla defines a `notecard` class that enforces colors on their inner
links. The color chosen is almost black and this makes any inner links
almost impossible to read.

This patch addresses the color issue and reapplies the same colors used
for other links.